Create separate installer to test UAC
If UAC is turned off the driver installation will always fail and
keep spawning processes. The only found solution was to check if
UAC is active but to do that the installer must request admin
permissions, therefore it should be separated from the main
installer. If this separate installer return the error `0` it
means UAC is on, otherwise skip driver install.